refactor: usecountdown into core hook
