+++ packages/core-lib/src/hooks/use-countdown.ts
@@ -1,0 +1,39 @@
+import { useEffect, useState } from 'react';
+
+interface DateProps {
+  date: string;
+}
+
+const getReturnValues = (countDown: any) => {
+  // calculate time left
+  const days = Math.floor(countDown / (1000 * 60 * 60 * 24));
+  const hours = Math.floor(
+    (countDown % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60)
+  );
+  const minutes = Math.floor((countDown % (1000 * 60 * 60)) / (1000 * 60));
+  const seconds = Math.floor((countDown % (1000 * 60)) / 1000);
+
+  const expired = days + hours + minutes + seconds <= 0;
+
+  return { days, hours, minutes, seconds, expired };
+};
+
+const useCountdown = (props: DateProps) => {
+  const countDownDate = new Date(parseInt(props.date)).getTime();
+
+  const [countDown, setCountDown] = useState(
+    countDownDate - new Date().getTime()
+  );
+
+  useEffect(() => {
+    const interval = setInterval(() => {
+      setCountDown(countDownDate - new Date().getTime());
+    }, 1000);
+
+    return () => clearInterval(interval);
+  }, [countDownDate]);
+
+  return getReturnValues(countDown);
+};
+
+export default useCountdown;
