+++ packages/ui-lib/src/base/countdown/CountDown.tsx
@@ -1,39 +1,8 @@
-import React, { useEffect, useState } from 'react';
+import useCountdown from '@bibliotheca-dao/core-lib/hooks/use-countdown';
 
 interface DateProps {
   date: string;
 }
-
-const useCountdown = (props: DateProps) => {
-  const countDownDate = new Date(parseInt(props.date)).getTime();
-
-  const [countDown, setCountDown] = useState(
-    countDownDate - new Date().getTime()
-  );
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setCountDown(countDownDate - new Date().getTime());
-    }, 1000);
-
-    return () => clearInterval(interval);
-  }, [countDownDate]);
-
-  return getReturnValues(countDown);
-};
-
-const getReturnValues = (countDown: any) => {
-  // calculate time left
-  const days = Math.floor(countDown / (1000 * 60 * 60 * 24));
-  const hours = Math.floor(
-    (countDown % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60)
-  );
-  const minutes = Math.floor((countDown % (1000 * 60 * 60)) / (1000 * 60));
-  const seconds = Math.floor((countDown % (1000 * 60)) / 1000);
-
-  return [days, hours, minutes, seconds];
-};
-
 interface TimeDisplay {
   value: any;
   type: any;
@@ -85,7 +54,7 @@ const ShowCounter = (props: Counter) => {
 };
 
 export const CountdownTimer = (props: DateProps) => {
-  const [days, hours, minutes, seconds] = useCountdown(props);
+  const { days, hours, minutes, seconds } = useCountdown(props);
   if (days + hours + minutes + seconds <= 0) {
     return <ExpiredNotice />;
   } else {
